+++ kogito-quarkus-examples/process-business-calendar-example/src/main/resources/calendar.properties
@@ -4,3 +4,9 @@ business.start.hour=0
 business.holiday.date.format=yyyy-MM-dd
 business.days.per.week =7
 business.cal.timezone= system default timezone..
+#business.end.hour=9
+#business.hours.per.day=1
+#business.start.hour=8
+#business.holiday.date.format=yyyy-MM-dd
+#business.days.per.week =1
+#business.cal.timezone= system default timezone..
